+++ pass-core-doi-service/src/main/java/org/eclipse/pass/doi/service/ExternalDoiServiceConnector.java
@@ -16,15 +16,10 @@
  */
 package org.eclipse.pass.doi.service;
 
-import static java.lang.Thread.sleep;
 import static java.util.concurrent.TimeUnit.SECONDS;
 
 import java.io.IOException;
 import java.io.StringReader;
-import java.util.HashSet;
-import java.util.Set;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import javax.json.Json;
 import javax.json.JsonObject;
 import javax.json.JsonReader;
@@ -44,60 +39,12 @@ public class ExternalDoiServiceConnector {
 
     private OkHttpClient client;
 
-    private Set<String> activeJobs = new HashSet<>();
-
     ExternalDoiServiceConnector() {
         OkHttpClient.Builder builder = new OkHttpClient.Builder();
         builder.connectTimeout(30, SECONDS);
         builder.readTimeout(30, SECONDS);
         builder.writeTimeout(30, SECONDS);
         this.client = builder.build();
-    }
-
-    /**
-     * check to see whether supplied DOI is in valid format after splitting off a possible prefix
-     *
-     * @return the valid suffix, or null if invalid
-     */
-    String verify(String doi) {
-        LOG.debug("Verifying doi format for " + doi );
-        if (doi == null) {
-            return null;
-        }
-        String criterion = "doi.org/";
-        int i = doi.indexOf(criterion);
-        String suffix = i >= 0 ? doi.substring(i + criterion.length()) : doi;
-
-        Pattern pattern = Pattern.compile("^10\\.\\d{4,9}/[-._;()/:a-zA-Z0-9]+$");
-
-        Matcher matcher = pattern.matcher(suffix);
-        return matcher.matches() ? suffix : null;
-    }
-
-    /**
-     * this simply protects the crossref service from a person hammering on a request thinking
-     * it wasn't processed, when it really is just slow coming back
-     *
-     * @param doi the doi to check active
-     * @return whether the doi lookup is still active
-     */
-    boolean isAlreadyActive(String doi) {
-        //stage 2: check cache map for existence of doi
-        //put doi on map if absent
-        LOG.debug("Checking to see if doi " + doi + " is already in process");
-        if (activeJobs.contains(doi)) {
-            return true;
-        } else {
-            // this DOI is not actively being processed
-            // let's temporarily prohibit new requests for this DOI
-            activeJobs.add(doi);
-            //longest time we expect it should take to create a Journal object, in
-            //milliseconds
-            int cachePeriod = 30000;
-            Thread t = new Thread(new ExpiringLock(doi, cachePeriod));
-            t.start();
-        }
-        return false;
     }
 
     /**
@@ -112,8 +59,9 @@ public class ExternalDoiServiceConnector {
         HttpUrl.Builder urlBuilder = HttpUrl.parse(service.baseUrl() + doi).newBuilder();
 
         if ( service.parameterMap() != null ) {
-            for ( String key : service.parameterMap().keySet() )
-            urlBuilder.addQueryParameter(key, service.parameterMap().get(key));
+            for ( String key : service.parameterMap().keySet() ) {
+                urlBuilder.addQueryParameter(key, service.parameterMap().get(key));
+            }
         }
 
         String url = urlBuilder.build().toString();
@@ -136,7 +84,8 @@ public class ExternalDoiServiceConnector {
             metadataJsonObject = reader.readObject();
             reader.close();
 
-            activeJobs.remove(doi);
+            service.unlockDoi(doi);
+
             return metadataJsonObject;
         } catch (JsonParsingException e) {
             if (responseString != null) {
@@ -150,27 +99,5 @@ public class ExternalDoiServiceConnector {
         return null;
     }
 
-    /**
-     * A class to manage locking so that an active process for a DOI will finish executing before
-     * another one begins
-     */
-    public class ExpiringLock implements Runnable {
-        private String key;
-        private int duration;
-
-        ExpiringLock(String key, int duration) {
-            this.key = key;
-            this.duration = duration;
-        }
-
-        public void run() {
-            try {
-                sleep(duration);
-                activeJobs.remove(key);
-            } catch (InterruptedException e) {
-                activeJobs.remove(key);
-            }
-        }
-    }
 }
